extracted program type and fetchUserProgram from workouts
