+++ frontend/src/components/Workouts.tsx
@@ -1,57 +1,27 @@
 import React, { useState, useEffect } from 'react'
-import NavigationButton from './NavigationButton'
-import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
+import Box from '@mui/material/Box'
 import { useAuth0 } from '@auth0/auth0-react'
-
-interface Program {
-  _id: string
-  programName: string
-  exercises: [
-    {
-      exerciseName: string
-      sets: number
-      reps: number
-      rest: string
-      weight?: number
-    }
-  ]
-  createdBy: string
-}
+import type Program from '../types/Program'
+import { fetchUserPrograms } from '../utils/fetchUserPrograms'
+import NavigationButton from './NavigationButton'
 
 const Workouts: React.FC = () => {
   const [programs, setPrograms] = useState<Program[]>([])
   const { user, getAccessTokenSilently } = useAuth0()
 
   useEffect(() => {
-    if (user === null || user === undefined) {
-      return
-    }
-
-    const fetchUserId = async (): Promise<void> => {
-      const accessToken = await getAccessTokenSilently()
-
+    const fetchData = async (): Promise<void> => {
       try {
-        const res = await fetch(`/programs/user/${user.sub}`, {
-          method: 'GET',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${accessToken}`
-          }
-        })
+        const accessToken = await getAccessTokenSilently()
 
-        if (!res.ok) {
-          throw new Error('Failed to fetch user')
-        }
-
-        const data = await res.json()
-        setPrograms(data.data)
+        const fetchedPrograms = await fetchUserPrograms(user?.sub, accessToken)
+        setPrograms(fetchedPrograms)
       } catch (error) {
-        console.log('error: ', error)
+        console.error('Error fetching programs:', error)
       }
     }
-
-    fetchUserId().catch((error) => {
+    fetchData().catch((error) => {
       console.error('Error fetching programs:', error)
     })
   }, [user, getAccessTokenSilently])
